Fixed launch loading issue, Add Task export
Add weekend to week view (Tasks can be scheduled on weekend so they
should now be grouped on it, for now there is no special style, but it
will come before release)
Finally fixed launch loading issue, where when loaded, the week view
would be empty.
Export to ical with tasks and notifications! Unfortunately notification
seem to not work on google calendar. will try to find if anything is
wrong in the code, but in outlook it works.